% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{babelwhale}
\alias{babelwhale}
\alias{babelwhale-package}
\title{Talking to both Docker and Singularity containers from R}
\description{
Talking to both Docker and Singularity containers from R
}
