% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dockerfile_to_singularityrecipe.R
\name{convert_dockerfile_to_singularityrecipe}
\alias{convert_dockerfile_to_singularityrecipe}
\title{Convert a Dockerfile to a Singularity recipe}
\usage{
convert_dockerfile_to_singularityrecipe(dockerfile, singularityrecipe,
  bootstrap = c("shub", "docker"), header = TRUE)
}
\arguments{
\item{dockerfile}{The path of the Dockerfile}

\item{singularityrecipe}{The path of the Singularity recipe to be created}

\item{bootstrap}{Where to bootstrap the Singularity image from}

\item{header}{Add a header to the Singularity recipe}
}
\description{
Convert a Dockerfile to a Singularity recipe
}
