% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_file.R
\name{copy_file}
\alias{copy_file}
\title{Copy a file from a container to the host system}
\usage{
copy_file(container_id, path_container, path_local)
}
\arguments{
\item{container_id}{The name of the container, usually the repository name on dockerhub.}

\item{path_container}{The path of the file inside the container}

\item{path_local}{The path of the file on the host system}
}
\description{
Copy a file from a container to the host system
}
\examples{
if (test_docker_installation()) {
  set_default_config(create_docker_config(), permanent = FALSE)
  copy_file("alpine", "/bin/date", tempfile())
}
}
