% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_create.R, R/config_default.R
\name{create_config}
\alias{create_config}
\alias{create_docker_config}
\alias{create_singularity_config}
\alias{get_default_config}
\alias{set_default_config}
\title{Backend configuration for containerisation}
\usage{
create_config(
  backend =
    get_env_or_null("BABELWHALE_BACKEND") \%||\%
    detect_backend()
)

create_docker_config()

create_singularity_config(
  cache_dir =
    get_env_or_null("SINGULARITY_CACHEDIR") \%||\%
    "./",
  use_cache = cache_dir != "./"
)

get_default_config()

set_default_config(config, permanent = TRUE)
}
\arguments{
\item{backend}{Which backend to use. Can be either \code{"docker"} or \code{"singularity"}.}

\item{cache_dir}{A folder in which to store the singularity images. Each TI method will require about 1GB of space.}

\item{use_cache}{If the singularity images are not cached, they will need to be downloaded every time a method is run.}

\item{config}{A config to save as default.}

\item{permanent}{Whether or not to save the config file permanently}
}
\description{
It is advised to define the \code{"BABELWHALE_BACKEND"} environment variable as \code{"docker"} or \code{"singularity"}.

When using singularity, also define the \code{"SINGULARITY_CACHEDIR"} environment variable,
which is the folder where the singularity images will be cached.
Each TI method will require about 1GB of space.

Alternatively, you can create a config and save it using \code{set_default_config()}.
}
\examples{
\dontrun{
config <- create_docker_config()
set_default_config(config)

config <- create_singularity_config(
  cache_dir = "~/babelwhale_singularity_images/"
)
set_default_config(config)
}

}
