% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Read a file from a container}
\usage{
read_file(container_id, path_container)
}
\arguments{
\item{container_id}{The name of the container, usually the repository name on dockerhub or singularityhub}

\item{path_container}{The path of the file inside the container}
}
\description{
Read a file from a container
}
