% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singularity_image_path.R
\name{singularity_image_path}
\alias{singularity_image_path}
\title{Determine the cached path of singularity images}
\usage{
singularity_image_path(container_id)
}
\arguments{
\item{container_id}{The name of the container, usually the repository name on dockerhub or singularityhub}
}
\description{
Determine the cached path of singularity images
}
