% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_docker_installation.R
\name{test_docker_installation}
\alias{test_docker_installation}
\title{Tests whether docker is correctly installed and available}
\usage{
test_docker_installation(detailed = FALSE)
}
\arguments{
\item{detailed}{Whether top do a detailed check}
}
\description{
Tests whether docker is correctly installed and available
}
