% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_singularity_installation.R
\name{test_singularity_installation}
\alias{test_singularity_installation}
\title{Tests whether singularity is correctly installed and available}
\usage{
test_singularity_installation(detailed = FALSE)
}
\arguments{
\item{detailed}{Whether top do a detailed check}
}
\description{
Tests whether singularity is correctly installed and available
}
\examples{
test_singularity_installation()

if (test_singularity_installation()) {
  test_singularity_installation(detailed = TRUE)
}

}
