% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{babelwhale}
\alias{babelwhale}
\title{Talking to both Docker and Singularity containers from R}
\description{
Provides a unified interface to interact with
docker' and 'singularity' containers.  You can execute a command
inside a container, mount a volume or copy a file.
}
