% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_create.R, R/config_default.R
\name{create_config}
\alias{create_config}
\alias{create_docker_config}
\alias{create_singularity_config}
\alias{get_default_config}
\alias{set_default_config}
\title{Backend configuration for containerisation}
\usage{
create_config(
  backend =
    get_env_or_null("BABELWHALE_BACKEND") \%||\%
    "docker",
  ...
)

create_docker_config()

create_singularity_config(
  cache_dir =
    get_env_or_null("SINGULARITY_CACHEDIR") \%||\%
    "./",
  use_cache = cache_dir != "./"
)

get_default_config()

set_default_config(config, permanent = TRUE)
}
\arguments{
\item{backend}{Which backend to use. Can be either `"docker"` or `"singularity"`.
If the value of this parameter is not passed explicitly, the default is defined by
the `"BABELWHALE_BACKEND"` environment variable, or otherwise just `"docker"`.}

\item{...}{Parameters to pass to `create_docker_config()` or `create_singularity_config()`.}

\item{cache_dir}{A folder in which to store the singularity images. Each TI method will require about 1GB of space.}

\item{config}{A config to save as default.}

\item{permanent}{Whether or not to save the config file permanently}

\item{prebuild}{If the singularity images are not prebuilt, they will need to be built every time a method is run.}
}
\description{
Backend configuration for containerisation
}
\examples{
\dontrun{
config <- create_docker_config()
set_default_config(config)

config <- create_singularity_config(
  cache_dir = "~/babelwhale_singularity_images/"
)
set_default_config(config)
}

}
