% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_docker_images.R
\name{list_docker_images}
\alias{list_docker_images}
\title{List docker containers}
\usage{
list_docker_images(container_id = NULL)
}
\arguments{
\item{container_id}{An optional container id}
}
\description{
List docker containers
}
\examples{
if (test_docker_installation()) {
  set_default_config(create_docker_config(), permanent = FALSE)
  list_docker_images()
}

}
