% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_image.R
\name{pull_container}
\alias{pull_container}
\title{Pull a container from dockerhub}
\usage{
pull_container(container_id)
}
\arguments{
\item{container_id}{The name of the container, usually the repository name on dockerhub or singularityhub}
}
\description{
Pull a container from dockerhub
}
