% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run a containerised command, and wait until finished}
\usage{
run(container_id, command, args = NULL, volumes = NULL,
  workspace = NULL, environment_variables = NULL, debug = FALSE,
  verbose = FALSE)
}
\arguments{
\item{container_id}{The name of the container, usually the repository name on dockerhub or singularityhub}

\item{command}{Character scalar, the command to run.}

\item{args}{Character vector, arguments to the command.}

\item{volumes}{Which volumes to be mounted}

\item{workspace}{Which working directory to run the command in.}

\item{environment_variables}{A character vector of environment variables. Format: \code{c("ENVVAR=VALUE")}.}

\item{debug}{If \code{TRUE}, a command will be printed for the user to execute.}

\item{verbose}{Whether or not to print output}
}
\description{
Run a containerised command, and wait until finished
}
